Refactor: Improve code structure and component modularity
