+++ components/layout/app-footer.tsx
@@ -1,0 +1,9 @@
+export const AppFooter = () => {
+  return (
+    <footer className="border-t mt-12">
+      <div className="container mx-auto px-4 py-6 text-center text-sm text-muted-foreground">
+        <p>Your API key and data never leave your browser. All storage is local.</p>
+      </div>
+    </footer>
+  );
+};
